pyproject.toml
[project]
name = "open-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.0.11",
    "pydantic>=2.11.3",
    "supabase>=2.15.1",
]
